Refactor and enhance Hanji SDK trading infrastructure
- Update package references to latest versions

- Add new ABI classes for Lob and LobFactory interactions

- Implement comprehensive trader interfaces and configurations

- Enhance price and decimal handling utilities

- Improve error logging and transaction management

- Add support for batch order operations and native token handling

- Restructure trader implementation with more robust state management

+++ Hanji.Sdk/Common/UserOrderExtensions.cs
@@ -34,18 +34,18 @@ namespace Hanji.Common
             };
 
             return new Order(
-                OrderId:     o.OrderId,
-                Price:       decimal.Parse(o.Price).FromHanjiPrice(pricePrecision),
-                Qty:         qty,
-                LeaveQty:    leaveQty,
-                ClaimedQty:  claimedQty,
-                Side:        side,
-                Symbol:      symbol,
-                Status:      status,
-                Type:        OrderType.Return,
-                Created:     DateTimeOffset.FromUnixTimeMilliseconds(o.CreatedAt),
+                OrderId: o.OrderId,
+                Price: ulong.Parse(o.Price).FromHanjiPrice(pricePrecision),
+                Qty: qty,
+                LeaveQty: leaveQty,
+                ClaimedQty: claimedQty,
+                Side: side,
+                Symbol: symbol,
+                Status: status,
+                Type: OrderType.Return,
+                Created: DateTimeOffset.FromUnixTimeMilliseconds(o.CreatedAt),
                 LastChanged: DateTimeOffset.FromUnixTimeMilliseconds(o.LastTouched),
-                TxnHash:     o.TxnHash);
+                TxnHash: o.TxnHash);
         }
     }
 }
